Developing frontends

+++ SunsOutBunsOut/Repositories/IBurgerRepository.cs
@@ -5,11 +5,11 @@ namespace Repositories
     public interface IBurgerRepository
     {
         Task<IEnumerable<Burger>> GetAllBurgersAsync();
-        Task<Burger> GetBurgerByIdAsync(long id);
+        Task<Burger> GetBurgerByIdAsync(int id);
         Task AddBurgerAsync(Burger Burger);
         Task UpdateBurgerAsync(Burger Burger);
-        Task DeleteBurgerAsync(long id);
-        Task<bool> BurgerExistsAsync(long id);
+        Task DeleteBurgerAsync(int id);
+        Task<bool> BurgerExistsAsync(int id);
         Task BulkAddBurgersAsync(IEnumerable<Burger> Burgers);
     }
 }
